marketplace/umaa/requirements.txt
Django==1.7.8
MySQL-python==1.2.5
djangorestframework==3.1.3
djangorestframework-jwt>=1.2.0
django-countries>=3.1.1
django-dynamic-initial-data>=0.4.1
PyYAML>=3.11
requests>=2.7.0